+++ src/libzyzzyva/WordTableModel.h
@@ -64,14 +64,14 @@ class WordTableModel : public QAbstractTableModel
         bool getFrontParentHook() const { return frontParentHook; }
         bool getBackParentHook() const { return backParentHook; }
         int getProbabilityOrder() const { return probabilityOrder; }
-        qint64 getPlayabilityValue() const { return playabilityValue; }
+        double getPlayabilityValue() const { return playabilityValue; }
         int getPlayabilityOrder() const { return playabilityOrder; }
         QString getLexiconSymbols() const { return lexiconSymbols; }
         void setWord(const QString& w) { word = w; }
         void setType(WordType t) { type = t; }
         void setWildcard(const QString& w) { wildcard = w; }
         void setProbabilityOrder(int p);
-        void setPlayabilityValue(qint64 p);
+        void setPlayabilityValue(double p);
         void setPlayabilityOrder(int p);
         void setLexiconSymbols(const QString& s);
 
@@ -99,7 +99,7 @@ class WordTableModel : public QAbstractTableModel
         QString word;
         WordType type;
         int probabilityOrder;
-        qint64 playabilityValue;
+        double playabilityValue;
         int playabilityOrder;
         QString wildcard;
         QString frontHooks;
